requirements.txt
PyGithub==1.26.0
python-gitlab==0.9.2
requests==2.8.1
six==1.10.0
wheel==0.24.0
